pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xlstruct"
version = "0.1.0"
description = "LLM-powered Excel parser — define a Pydantic schema, get structured data from any Excel file"
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "DanMeon" }]
keywords = ["excel", "llm", "pydantic", "structured-data", "extraction", "spreadsheet"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Spreadsheet",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "openpyxl>=3.1",
    "python-calamine>=0.6.2",
    "instructor>=1.14.5,<2",
    "fsspec>=2024.1",
    "typer>=0.9",
    "tiktoken>=0.7",
]

[project.urls]
Homepage = "https://github.com/DanMeon/xlstruct"
Repository = "https://github.com/DanMeon/xlstruct"
Issues = "https://github.com/DanMeon/xlstruct/issues"
Changelog = "https://github.com/DanMeon/xlstruct/blob/main/CHANGELOG.md"

[project.optional-dependencies]
s3 = ["s3fs"]
azure = ["adlfs"]
gcs = ["gcsfs"]
openai = ["openai"]
anthropic = ["anthropic"]
gemini = ["google-genai"]
all-storage = ["s3fs", "adlfs", "gcsfs"]
all-llm = ["openai", "anthropic", "google-genai"]
all = ["s3fs", "adlfs", "gcsfs", "openai", "anthropic", "google-genai"]
docker = [
    "aiodocker>=0.26.0",
]

[project.scripts]
xlstruct = "xlstruct.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/xlstruct"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
show_error_codes = true

[[tool.mypy.overrides]]
module = ["openpyxl.*", "fsspec.*", "calamine.*", "anthropic.*", "instructor.*", "tiktoken.*", "aiodocker.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "xlstruct.codegen.backends.docker"
disallow_untyped_calls = false

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "adlfs>=2026.2.0",
    "azure-storage-blob>=12.28.0",
    "gcsfs>=2026.2.0",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=6.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.15.4",
    "s3fs>=2026.2.0",
    "xlwt>=1.3.0",
]
